refactor analytics diagnostics policy
