+++ src/app/routers/analytics_diagnostics.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import logging
-import re
 from typing import Annotated
 
 from fastapi import APIRouter, Header, HTTPException, Path, status
@@ -11,24 +10,18 @@ from app.contracts.analytics_diagnostics import (
     AnalyticsDiagnosticsResponse,
 )
 from app.observability.analytics_ui import (
-    ANALYTICS_UI_FORBIDDEN_FIELDS,
     emit_gateway_protected_diagnostics_audit_log,
+)
+from app.services.analytics_diagnostics_service import (
+    is_authorized_operator_role,
+    is_safe_support_reference,
+    resolve_support_reference,
 )
 from app.services.caller_context import caller_context_headers
 
 logger = logging.getLogger("analytics_ui.gateway")
 
 router = APIRouter(prefix="/api/v1/analytics-ui", tags=["Analytics Diagnostics"])
-
-_SAFE_SUPPORT_REFERENCE = re.compile(r"^gdiag-[A-Za-z0-9][A-Za-z0-9_.:-]{1,121}$")
-_AUTHORIZED_OPERATOR_ROLES = {
-    "admin",
-    "ops",
-    "operations",
-    "operator",
-    "support",
-    "support_operator",
-}
 
 
 @router.get(
@@ -84,8 +77,7 @@ async def lookup_analytics_diagnostics(
         )
         raise
 
-    normalized_role = (role or "").strip().lower().replace("-", "_")
-    if normalized_role not in _AUTHORIZED_OPERATOR_ROLES:
+    if not is_authorized_operator_role(role):
         emit_gateway_protected_diagnostics_audit_log(
             logger=logger,
             status_code=status.HTTP_403_FORBIDDEN,
@@ -99,7 +91,7 @@ async def lookup_analytics_diagnostics(
             },
         )
 
-    if not _SAFE_SUPPORT_REFERENCE.fullmatch(support_reference):
+    if not is_safe_support_reference(support_reference):
         emit_gateway_protected_diagnostics_audit_log(
             logger=logger,
             status_code=status.HTTP_400_BAD_REQUEST,
@@ -113,79 +105,10 @@ async def lookup_analytics_diagnostics(
             },
         )
 
-    response = _resolve_support_reference(support_reference)
+    response = resolve_support_reference(support_reference)
     emit_gateway_protected_diagnostics_audit_log(
         logger=logger,
         status_code=status.HTTP_200_OK,
         reason="lookup_succeeded",
     )
     return response
-
-
-def _resolve_support_reference(support_reference: str) -> AnalyticsDiagnosticsResponse:
-    normalized = support_reference.lower()
-    if "risk" in normalized:
-        panel = "risk-summary"
-        operation = "analytics.risk.calculate"
-        service = "lotus-risk"
-    elif "performance" in normalized:
-        panel = "performance-summary"
-        operation = "performance.workspace-summary"
-        service = "lotus-performance"
-    else:
-        panel = "unknown"
-        operation = "analytics-ui.lookup"
-        service = "unknown"
-
-    if "permission" in normalized or "denied" in normalized or "blocked" in normalized:
-        supportability_state = "permission_blocked"
-        reason = "upstream_authorization_denied"
-    elif "degraded" in normalized or "partial" in normalized:
-        supportability_state = "degraded"
-        reason = "upstream_partial_or_degraded"
-    else:
-        supportability_state = "ready"
-        reason = "safe_reference_resolved"
-
-    return AnalyticsDiagnosticsResponse(
-        supportReference=support_reference,
-        panel=panel,
-        supportabilityState=supportability_state,
-        safeDimensions={
-            "operation": operation,
-            "service": service,
-            "state": supportability_state,
-            "reason": reason,
-        },
-        operatorGuidance=_operator_guidance(supportability_state),
-        forbiddenFields=sorted(ANALYTICS_UI_FORBIDDEN_FIELDS),
-    )
-
-
-def _operator_guidance(supportability_state: str) -> list[str]:
-    if supportability_state == "permission_blocked":
-        return [
-            "Confirm caller entitlement and region posture in the authoritative upstream service.",
-            (
-                "Use protected operational evidence for correlation lookup; do not add identifiers "
-                "to dashboard labels."
-            ),
-        ]
-    if supportability_state == "degraded":
-        return [
-            (
-                "Check upstream service health and partial-failure posture for the safe operation "
-                "dimension."
-            ),
-            "Escalate with bounded service, operation, state, and reason only.",
-        ]
-    return [
-        (
-            "Use the safe dimensions to locate the relevant analytics panel and upstream service "
-            "posture."
-        ),
-        (
-            "Keep raw request, response, trace, correlation, and client identifiers out of "
-            "analytics UI telemetry."
-        ),
-    ]
